% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrect.R
\name{biasCorrect}
\alias{biasCorrect}
\title{Performs inhomogeneity correction (IHC) on an array of images.}
\usage{
biasCorrect(image, mask, splineParam = 2000,
  mc.cores = getOption("mc.cores", 2L), ...)
}
\arguments{
\item{image}{Required character vector of image locations or raster object.}

\item{mask}{Required character vector of image location or raster object of the tissue mask.}

\item{splineParam}{Controls smoothness of bias field. Argument passed to ANTsRCore::n4BiasFieldCorrection.}

\item{mc.cores}{Number of cores to use for parallel things.}

\item{...}{Other arguments passed to ANTsRCore::n4BiasFieldCorrection.}
}
\value{
Returns the bias corrected images as a raster stack object.
}
\description{
Adjusts for slow frequency fluctuations in image intensities, i.e. field bias in MRI.
This function computes a k component gamma mixture model marginally for each channel and returns a mixfit object to perform image normalization using the XX function.
}
\keyword{bias}
\keyword{correction}
