% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmix.R
\name{fitmix}
\alias{fitmix}
\title{Fits a mixture model on an array of images}
\usage{
fitmix(image, mask, lambda = NULL, k = 3, alpha = c(0.333, 3.33,
  33.3), beta = c(c(0.222, 2.22, 22.2)),
  mc.cores = getOption("mc.cores", 2L), subsamp = 1, maxpix = NULL,
  ...)
}
\arguments{
\item{image}{Required character vector of image locations or raster object.}

\item{mask}{Required character vector of image location or raster object of the tissue mask.}

\item{lambda}{k length vector of mixing proportions.}

\item{k}{number of mixture components. Can be thought of as the number of tissue classes distinguishable by this channel. Probably better to have too many than too few. With too too many convergence issues may occur.}

\item{alpha}{k length vector of shape parameters}

\item{beta}{k length vector of beta parameters}

\item{mc.cores}{Number of cores for parallel things. With large images and a large number of cores this may create memory issues. With many channels, it may be better to distributed computing over a cluster.}

\item{subsamp}{Proportion of pixels within mask to use to estimate mixture model. gammamixEM can fail with a large number of pixels.}

\item{maxpix}{Maximum number of pixels to use to estimate mixture model.}

\item{...}{Other arguments passed to gammamixEM}
}
\value{
Returns a mixfit object with the following values:
}
\description{
This function computes a k component gamma mixture model marginally for each channel and returns a mixfit object to perform image normalization using the XX function.
}
\keyword{gamma}
\keyword{mixture}
\keyword{model}
